refactor (tests): move shared test code to ... shared.py
